Fix missing cached properties in BlueZ backend
BlueZ will cache properties of devices it has seen before.  By only relying on the PropertiesChanged signal, we only get properties that are changing, like RSSI and miss properties that don't change.

To fix this, we also save all properties of all devices that are returned by GetManagedObjects. Then, when we receive a PropertiesChanged singal indicating that a device is active, we merge the cached properties with the new properties for that device.

GetManagedObjects need to be moved after addMatch to avoid the race condition where properties change between getting properties from GetManagedObjects and registering the callback for PropertiesChanged.