fix characteristic write response parameter on Windows
In the Windows API, the only difference between WriteValueWithResultAsync and WriteValueAsync is the return type. When the GattWriteOption parameter is omitted, both default to GattWriteOption.WriteWithResponse. To actually write without response, we need to make use of that parameter.

Keeping WriteValueWithResultAsync instead of WriteValueAsync since it returns more information that could be useful for debugging.