Fix leaking of callbacks from DBus on multiple calls to discover()
If the rules are only added to Bluez via DBus but never removed again,
the calls pile up if discover() is called multiple times in one session.
This is especially harmful for long-running applications that connect
and disconnect to the same or different servers over and over. The time
to await discovery() becomes very long (up to minutes) even with smaller
timeouts.

Make sure the callbacks registered are cleaned up before discover() is
left.