Consolidate set_disconnected_callback() implementations
All backends had the same implementation for set_disconnected_callback()
so we can have a common implementation in the abstract base class instead.

This also allows us to pass the disconnected callback as a keyword
argument when the BleakClient object is created. This can help avoid
a race condition where a device might disconnect before the callback
can be set.

The example program is updated accordingly to avoid such a race condition.

The existing set_disconnected_callback() can also be used at any time to
change or remove the current callback.