consolidate BleakGATTService.get_characteristic()
All backends had the same implementation, so it can be moved to the
abstract base class.

BleakGATTService.uuid properties are modified so that all platforms
return a lower case xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx string.

The filter is modified so that if a string argument is given, then it
will be converted to lower case to match the uuid property.