corebluetooth: use threading.Event to wait for centralManagerDidUpdateState_
Previously, we were using an `asyncio.Event` to wait for the initial
centralManagerDidUpdateState_ callback. However we were not always
waiting for this before using the `CBCentralManager` in all code paths.

By using `threading.Event` instead, we can do a (short) blocking wait
in the initializer to ensure that `CBCentralManager` is ready before
we do anything else.