Removed separate implementation of global ``discover`` method.
The "old" `discover()` method was largely duplicating the code from `BleackScanner.discover()`, with the exception of the Mac backend which looked like it might not even work correctly and ignored kwargs.

This removes the `discover()` method from the documentation and removes the separate implementation. The name is still kept and made as a synonym for for `BleakScanner.discover` so that we don't completely break existing code that might be using it.