Added filtering on advertisement data callbacks on BlueZ backend.
Currently BlueZ invoke an advertisement data callback every time the
RSSI changes while scanning. This is much more frequent than other
backends. To make things more consistent cross-platform, this adds some
filtering so that we only invoke the callback the first time the device
is seen or when a relevant property changes. This brings the behavior
more inline with the default macOS behavior. (Windows still needs work
to get the same behavior.)

Issue: #394