BlueZ: Fixed stop_notify() error logged on device disconnect.
This removes the call to stop_notify() in the cleanup code path in the
BlueZ backend. This was causing an error message to always be logged
when notifications were enabled and a device was disconnected.

As described in the code comment. This call is not necessary and can be
removed.

Fixes: #472

(No changelog since this bug was introduced in changes since the
previous release)