Add CI environment variable check to import.
Using Bleak on CI will fail if `bluetoothctl` is not installed. This
adds a check for the CI environment variable that is set by default in
many CI environments (e.g. GitHub Actions, Travis CI, etc.). Since we
aren't sure that the value will be the same on all CI platforms, we just
check for the presence of the variable. The check extends the same
workaround that is already in place for building on readthedocs.org.