Removed use of CMDConnectionState in CoreBluetooth backend.
This removes the peripheral state management from the central manager.
The CBPeripheral objects keep track of their own state which can be
fetched via the state() property, so we don't need to keep track of the
state separately.

The use of the state change event is replaced with futures. Also, this
change now properly waits for a callback after cancelling the connection
in the event of a asyncio.TimeoutError which happens to fix #538.