Fix connecting to multiple devices on Mac
This fixes connecting to multiple devices when using a single
BleakScanner in the CoreBluetooth backend. Previously, the central
manager held a reference to a single peripheral, so if a second
peripheral was connected using the same central manager, it would
overwrite the first.

This removes the coupling so that the peripheral is passed as an
argument to the central manager methods instead.