Fix handling of futures for same UUID on CoreBluetooth
This fixes potential deadlocks when using a service, characteristic or
descriptor with the same UUID on the CoreBluetooth backend. Previously,
the futures were stored by UUID for later lookup in callbacks. However,
since since the same UUID can be used multiple times, we could end up
with futures being replaced in a dictionary before they were completed,
which would result in a deadlock for the waiter.