CoreBluetooth scan start/stop improvements
This makes use of ObjC-style key-value observer to monitor the isScanning
property on CBCentralManager. This allows us to avoid the polling wait.

Previously, the wait was only implemented when stopping scanning, so it
is added to the start as well.

Also, the higher-level function is fixed to allow calling stop()
multiple times without raising an error. This is needed since the
stop() method can be called by __aexit__() even if the scanner is
already stopped.