winrt: refactor error handling
There was quite a bit of duplicated code checking the error status
returned from GATT methods in the WinRT backend. This is refactored
into a single function the check the status and raise an exception.

Error messages are changed to use the handle instead of uuid to identify
characteristics and descriptors.