winrt: remove get_services(use_cached) option
This removes the `use_cached` kwarg from `BleakClient.get_services()`
and the pass-through from `connect()` and `__init__()`. This was used
to work around issues where the cached values may be incorrect.

Since a large segment of our user base is DIY enthusiasts that trigger
this problem by changing the configuration of their DIY device, it makes
sense to just never use the cache and avoid the problem in the first
place instead of making users have to encounter the error, research it,
find the troubleshooting guide and apply the workaround.

If there is ever a legitimate use case for using cached values, we can
reconsider the feature.