CoreBluetooth: delete futures after awaiting
In the CoreBluetooth backend, futures are used to pass results/errors
from the delegate callbacks to methods that are waiting for them. The
futures are stored as part of the object state so that the delegates
can access them.

Prior to this change, completed futures were not removed from the
object status. This worked most of the time because a second call
of the same method would replace the completed future from the prior
call. However, in the case of notifications and reads, the same
delegate callback is shared with two methods. So when both of these
were used, notifications would attempt to complete the already completed
read future which raises an InvalidStateError.

Fixes #675.