CoreBluetooth: pass return values via futures
This updates CoreBluetooth to read the changed value in the delegate
callback and pass it back to the BleakClient method via the future
object.

By reading the value in the delegate callback rather than the method
scheduled by call_soon_threadsafe(), we ensure that we get the original
value from the delegate callback and not a possibly changed value later.