winrt: combine adv data and scan rsp data
This changes how advertising data is cached and combined. Previously,
we just cached scan response data to get the local name later. Separate
callbacks were called for advertising data and scan response data which
makes more work for users who are interested in filtering on data that
is split between the two packets.

Now, the advertising data and scan response data are both cached together
as a single object and this object is used to create the Bleak advertising
data structure so that the advertising data seen by Bleak users now
contains the information from both advertising data type combined.