corebluetooth: don't log scanner error in app
If bleak is running in a .app (e.g. created with py2app), then scanning
in macOS 12 works as it did in previous versions, so we don't need to
log the error. We can determine this by checking the bundle id since
Python itself is a .app and has a bundle id.

Related: https://github.com/hbldh/bleak/issues/720