Always call self._central_manager_delegate.disconnect
Previously this was only called if the peripheral wasn't already
disconnected.

This is a problem when devices might disconnect before you
are done with them.

It could lead to a leak where the handler was never disconnected
and you could end up with duplicate disconnect callbacks for
the same device.

The client objects themselves also leaked.

The existing code had the effect of masking a duplicate callback.
corebluetooth seems to send a notification with no error when
we disconnect the disconnect handler. So users previously only
saw disconnect events caused by errors (the device going away).

Now that we disconnect from the handler we get an extra notification.
To avoid an API break, we now only invoke the users callback if there
was an error condition. This simulates the old behaviour.