backends/scanner: implement scanning_mode arg in all backends
Previously, only the Windows backend supported this argument. This
adds the argument to all backends. An implementation is added to the
android backend. macOS doesn't support this, so an exception is raised
in that case. The linux backend will require extensive changes, so
support will be added in a later commit and raises NotImplementedError
for the time being.