bluez: use same callback for passive and active scan
This replaces the callback mechanism in BlueZ passive scanning with the
one from active scanning. This makes both scans work the same.

After some experimentation, we realized that the BlueZ advertisement
monitor APIs are really just for detecting if a device goes in range
or out of range, which isn't what we need in Bleak. So basically, we
are just using it as a hack to initiate passive scanning since there
isn't a flag in "StartDiscovery" or "SetDiscoveryFilter" to select
passive scanning.