winrt/client: add 3rd state for use_cached_services
This adds a 3rd state for the backend-specific use_cached_services
option. It can now be None, True or False.

If None, get_serv/char/desc_async() is called without any arguments
which is the new default so may change behavior (and hopefully fix
some issues). If True or False, the second form is called to force
the desired behavior.