bluezdbus/manager: call advertisement callbacks from InterfacesAdded
We have found that when doing passive scanning, some devices only send
one advertisement and then go to sleep for a while. BlueZ triggers
InterfacesAdded and InterfacesRemoved signals for these devices since
the sleep time is long enough for them to be considered no longer
present.

This caused advertisements to be missed since we were previously just
relying on PropertiesChanged signals with an RSSI change to determine
if a device is actually advertising.

By calling the callbacks on InterfacesAdded signals as well, we can
catch the very first and possibly only advertisement from these sorts
of devices.