backends/characteristic: add max_write_without_response_size property
This adds a new max_write_without_response_size property to
BleakGattClient that can be used to ensure data will be able to be used
in a characteristic write without response command without causing
an OS error.

The NUS example is updated to show how to make use of this property.