backends/bluezdbus/manager: add object lookup maps
This adds maps to improve performance when calling get_services().

Previously we had to enumerate all BlueZ D-Bus objects for each service,
characteristic and descriptor. When there were many devices, this was
a performance bottleneck.

Instead, we can create a map of the D-Bus object tree as new interfaces
appear, then use this map as an efficient way to find the child D-Bus
object paths in the GATT tree.

Fixes #927.