backends/bluezdbus/client: fix not disconnecting if get_services() throws
It is possible for get_services() to raise an exception, e.g. if
external code cancels the task. If this happens, we need to be sure
that the device disconnects so the BlueZClient is in a consistent
state (if the connect() method raises and exception, is is expected
that the device will not be connected).

To fix this, we can just move the call to get_services() inside of the
existing try block that will disconnect on any exception, then reraise
the exception.

Fixes #951.