backends/winrt: add connection check to BleakClient methods (#973)
This adds a connection check to all methods that require a connection in the WinRT backend similar to what is already done in the BlueZ backend.

This fixes an issue where we would get an AttribueError in the connect method while getting services caused by the device disconnecting after connecting but before getting the services. Now a more informative exception is raised instead.