backends/scanner: improve BLEDevice creation performance
This changes all scanner backends so that they only create a BLEDevice
object once per scan. For example, in several backends, a new BLEDevice
was created for each discovered device each time the discovered_devices
property was called, which could be significant if many (~100) devices
were discovered.

This also caches the last advertisement data along with the BLEDevice
which is expected to be used as an alternative to BLEDevice.metadata
in the future.