scanner: add discovered_devices_and_advertisement_data property
This adds a new discovered_devices_and_advertisement_data property
to BleakScanner that can be used to get the last advertisement data in
addition to the BLEDevice of the devices discovered during the scan
session.

A new argument is added to the BleakScanner.discover() class method to
allow this method to return the new property as well.