backends/bluezdbus/manager: replace startswith in hot path
Receiving advertisements is a hot path in the code. We can improve
performance by avoiding using startswith. In this case we conveniently
already have the adapter path available as a D-Bus property, so we
can use that instead.

Also drop comment while we are touching this since deep copy wouldn't
actually make much of a difference since the dictionary value should
be treated as immutable already.