bluez: Implement unpair() method
Operates by calling RemoveDevice on the Adapter object, which can be done
regardless of the device's connection status.

If the device is connected then calling this method will cause Bluez to
disconnect. Calling this method will also cause Bluez to delete any bonding
information and GATT attribute cache.

This seems like the only applicable way to unpair a device in Bluez. It's
possible to set the Trusted parameter (readwrite) on a Device object, but not
the Paired parameter.