BleakClient: add close() method
This adds a close() method to the BleakClient class and backends. This
is used to explicitly close an resource, like the D-Bus socket in the
BlueZ backend.

In the BlueZ backend, the D-Bus socket is no longer closed on
disconnect. This is needed to avoid problems with D-Bus methods failing
with EOFError() or hanging forever because the socket was disconnect.