winrt/client: fix TimeoutError for certain devices
Some devices will get a TimeoutError when trying to connect because
setting `self._session.maintain_connection = True` is not enough to
trigger a connection to the device. We also have to start an interaction
with the device. This can be done by calling `self.get_services()` which
we need to do here anyway. This also make `get_services()` be included
in the timeout, which is probably a good thing.

Fixes: #604