winrt/client: separate service cache mode
This separates the cache mode used for getting services from the one
used for getting characteristics and descriptors. Since we change the
cache mode on a retry, we want to be sure that only the service cache
mode is changed on the retry and not the characteristic or descriptor
cache modes.