uuids: remove invalid UTF-8
This removes some invalid UTF-8 characters. This was noticed when
Bleak examples were run with `python -m trace --trace`. I caused the
trace to crash with a UnicodeError. The Chinese name for Huawei has
been removed from the Bluetooth assigned numbers doc so we removed
all the Chinese names just to keep it simple.