winrt:client: disable retry on services changed
The change from 4756c1ed seems to be causing problems for users. It
appears that Windows doesn't handle cancellation of async methods
very well which causes hangs waiting for the cancelled methods to
return or the next method call hanging because the previous left the
Windows Bluetooth stack in a bad state.

This leaves the workaround in the code in case we still need it but
disables it for now. A WinRTClientArgs option could be added later
to optionally use it.