backends/scanner: add handle_early_stop hooks
All supported platforms provide some sort of hook that we can use to
get a callback when scanning stops. On Mac and Linux, there is only a
simple boolean value indicating if scanning is currently in progress.
Windows and Android provide additional error information, but this
is ignored for now since it isn't available cross-platform.

What we actually do with this hook is not yet implemented.