bluezdbus/manager: suppress org.bluez.Error.NotReady when stopping scanner
When stopping scanning, we don't want to raise errors if we can be sure
that scanning is in fact stopped.

`org.bluez.Error.NotReady` can be triggered if the Bluetooth adapter
is turned off after scanning started, in which case we know that it
is not scanning, so it is safe to ignore this error.