bgapi: explicitly close connections if we time out
Without this, I was occasionally seeing apparent bugs in the NCP
firmware, where it would start handing out connection handle 0, without
failure.  The api docs imply:

"If a connection can't be established, for example, the remote device
has gone out of range, has entered into deep sleep, or is not
advertising, the stack will try to connect forever. In this case, the
application will not get an event related to the connection request. To
recover from this situation, the application can implement a timeout and
call sl_bt_connection_close to cancel the connection request."

that we should be explicitly calling close therefore, and adding this
eliminated these errors in testing.

Signed-off-by: Karl Palsson <karlp@etactica.com>