winrt/scanner: drop waiting for matching scan rsp
This removes the logic that skips the detection callback when expecting
a scan response. It is valid for devices to broadcast ADV_IND but not
respond to a SCAN_REQ (there is no flag that means connectable but not
scannable so ADV_IND is the only choice in this case).

Fixes: https://github.com/hbldh/bleak/issues/1211