bleak: fix leaking of ensure_future()
As noted in the Python docs, a reference to the return value of
`asyncio.ensure_future()` must be held in order to prevent it from
being garbage collected before the task completes.

This applies the recommended fix from the docs of holding the reference
in a global set and then discarding the reference when the task
completes.

Also change `asyncio.ensure_future()` to `asyncio.create_task()` while
we are touching this since the minimum supported Python version is now
3.7.

Fixes: https://github.com/hbldh/bleak/issues/1258