winrt/client: ignore session closed event during connect
As described in https://github.com/hbldh/bleak/issues/1280, it is
possible for a session to be closed then activated again while getting
services. This will cause an OSError because the closed event triggers
the disconnect cleanup code, then trying to use the session after that
fails because it has been closed.

This works around the problem by disabling the disconnect cleanup code
until the connect method has returned successfully. If the connect
method raises an exception, it handles cleanup already so the cleanup
code in the session status event handler never gets called in this case.