backends/winrt/client: handle arbitrary access denied
We had a hard-coded list of services know to return an access denied
error when attempting to enumerate characteristics. However, since
we don't know Windows internals, the list was not exhaustive and it
could also change in the future.

Instead, we can just check for the access denied error and skip any
service that returns that error which should also handle additional
services we don't know about yet.

Fixes: https://github.com/hbldh/bleak/issues/1291