examples/two_devices: fix example
This fixes several issues with the implementation. Notably, on Linux
trying to scan and connect to two devices at the same time results in:

    bleak.exc.BleakDBusError: [org.bluez.Error.InProgress] Operation already in progress

This is worked around by using an asyncio.Lock to ensure only one scan
and connect is called at a time.

Fixes: https://github.com/hbldh/bleak/issues/1036

Also fixes error due to main() not being a coroutine:

    ValueError: a coroutine was expected, got <_GatheringFuture pending>

Fixes: https://github.com/hbldh/bleak/issues/1288

Finally, command line args are added to make it easy to run the example
without modifying the code.