backends/winrt: improve error messages with enums (#1370)
It has been noted that IntEnum.__str__() has changed in Python 3.11 and
now only prints integers instead of the name of the enum. This changes
all logging messages that use enums in the WinRT backend to use %r for
repr() so that we get both the name and the value.

Fixes: https://github.com/hbldh/bleak/issues/1284

Also changes uses of str.format() to f-strings and remove uses of
format string from loggers while we are touching this.