Fixed bleak recipe for android
The current develop branch of p4a uses the setup.py to install the recipe and doesn't seem compliant with PEP517 yet.

The temporary fix is to provide a setup.py during the build stage for android.

Also, the recipe is now more generic, so it can easily be used in your own projects.