backends/winrt: raise exception when trying to scan with STA
The pywin32-related packages will implicitly set the threading model to
STA when imported. Since Bleak is using WinRT for async methods and we
don't have a Windows event loop running, we need to let users know that
Bleak is not going to work in this case.

Also add a utility function and troubleshooting docs to provide a
workaround for this issue.

Fixes: https://github.com/hbldh/bleak/issues/1132