backends/scanner: always filter by service_uuids
It was noted that on Linux, if another app was scanning at the same
time, BlueZ would trigger RSSI changes for all devices, even if they
they didn't match the service_uuids filter.

This change ensures that we always filter by service_uuids, even if
the OS isn't doing it for us.

On Windows, the OS wasn't filtering for us anyway, so we can just move
that code to the shared call_detection_callbacks() so that all backends
will make use of it.

Fixes: https://github.com/hbldh/bleak/issues/1534