Provide a local definition of user32.SetTimer, user32.KillTimer and ole32.CoGetApartmentType on Windows
By overriding the type definition of SetTimer, we prevent None being passed as the callback parameter. This causes problems when integrating Bleak in larger Windows applications that call SetTimer in this way.
By defining the function type definition locally, we don't override the global definition so existing code that uses None as a callback keeps working.

To prevent other problems, we also make the definitions of KillTimer and CoGetApartmentType local.