Corebluetooth: add notification descriminator (#1742)
Add a new callback that can help discriminate between characteristic value changes due to a notification and ones due to a read response. CoreBluetooth uses the same delegate callback for both and doesn't have any info available in the parameters to know what is the difference between the two.

To work around this allow the user to provide a discriminator function that can inspect the value to determine if it was from a notification. This way, if a notification happens while we are waiting for a read response, we don't accidentally send the notification value as the read response value.