bleak: use backend-specific key for seen_devices
Replace use of address with backend-specific key for the seen_devices
dictionary.

As we found out, in the BlueZ backend, the address may not match the
D-Bus path in the case of private resolvable addresses. So to avoid
problems with converting the address to a D-Bus path, we need to use
the D-Bus path as the key in the seen_devices dictionary.

For good measure, we also use the macOS UUID for the device even when
using the use_bdaddr hack to get the actual address.

Other backends continue to use the address as the key, as there doesn't
appear to be any better alternative.