backends/device: restore **kwargs in BLEDevice constructor
The BLEDevice() constructor technically isn't considered part of the
public API, so a breaking change was made removing `rssi` and `**kwargs`.
However, there are users that implement their own backends, so effectively
this should be considered public and the breaking change reverted.

It looks like all users are passing `rssi` as a keyword argument, so
just restoring `**kwargs` should be sufficient.

Signed-off-by: David Lechner <david@pybricks.com>