backends/winrt/client: call disconnect callback on disconnect
Since commit 689f21e ("backends: winrt: don't wait for disconnect"), we
no longer wait for the disconnect event before returning from the
disconnect() method. However, some applications may be expecting the
disconnect callback to be called on disconnect. So we need to call the
disconnect callback manually in the disconnect() method. The event is
removed after this to prevent double calls. However, there is still the
possibility that the event handler could be in the asyncio event loop
queue, so we might need to reconsider how we handle this in the future.