bleak: new BleakBluetoothNotAvailableError exception
Add a new BleakBluetoothNotAvailableError specialized exception to be able to provide more useful error messages on how to recover from Bluetooth not being available. This exception includes a `reason` property that can be used to programatically check the reason as well.

This is only implemented in the corebluetooth backend currently. Other backends will be adapted in later commits.