backends: bluez: use AcquireNotify when possible
Using StartNotify in the BlueZ backend has a similar issue as the CoreBluetooth backend in that we get the same signal that the characteristic value has changed for both notifications and reads and there isn't a way to tell the difference.

Using AcquireNotify instead of StartNotify creates a pipe that only receives notifications for a single characteristic. Having this separate communication channel makes it known that the change was a notification and not a read. 

As a bonus, this should be more performant since it avoids a D-Bus signal on every notification (although we didn't actually measure it).