backends/corebluetooth: add 'NSNumber' type hint.
A 'NSNumber' can theoretically contain a float and not an int. So it is
better to convert it explicitly to an 'int'. This is can be seen by the
new type hint for 'NSNumber'.

Also convert rssi explicitly to 'int' since this value is part of the
public API to avoid any unexpected surprises.