tests: Run Integration tests for BlueZ in CI
Add test infrastructure to use Bumble + VHCI on Linux to test without any real Bluetooth hardware.

Also add tooling to create an Alpine VM with a VHCI to run the integration tests completely in CI. This is needed since the GitHub runners don't have Bluetooth enabled in the Linux kernel.