tests/integration: work around Bumble-Linux timeout
Add a hack to work around a Bumble missing feature that causes a timeout
in the Linux kernel.

Before this change, tests that made a connection to a peripheral would
have a 20 second delay before disconnecting. After this fix, the delay
is only 2 seconds (due to a different delay in bluetoothd).